requirements.txt
aioredis==2.0.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.1.3
certifi==2024.2.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
Faker==25.2.0
fastapi==0.111.0
fastapi-cli==0.0.3
fonttools==4.51.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
networkx==3.3
orjson==3.10.3
packaging==24.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
